Move new RPB share buttons in to the 'sort and filter your report' box alongside the export button (makes enough sense for them to be there, since it's mainly those options that get stored in the url/are shareable). Cleanup in a handful of spots touching the WriteToClipboard component
