+++ client/src/components/PresentationalPanel.scss
@@ -90,25 +90,18 @@
     width: 20px;
     height: 20px;
   }
-
   button.panel-heading-utils {
     background-color: Transparent;
     border: none;
     margin-right: 5px;
     padding: 0;
+
+    &:active {
+      transform: scale(1.15);
+    }
   }
 
 .panel-separator {
   margin: 20px 60px 20px 60px;
   border-top: 1px solid rgba($separatorColor, 0.5);
-}
-
-.panel-heading-utils:active {
-  transform: scale(1.15);
-}
-
-@media (hover: none){
-  .panel-heading-utils:hover, .panel-heading-utils:active {
-    transform: scale(1.15);
-  }
 }
