+++ client/src/rpb/simple_view.js
@@ -25,7 +25,8 @@ const granular_rpb_link_for_filter = (currentProps, filter) => rpb_link({ ...cur
 
 
 class SimpleView extends React.Component {
-  render(){
+
+  render() {
     const {
       subject,
       table,
@@ -44,7 +45,6 @@ class SimpleView extends React.Component {
       on_toggle_col_nick,
       on_toggle_deptBreakout,
     } = this.props;
-
     return (
       <div> 
         <LabeledBox label={<TextMaker text_key="blue_text_table_controls" />}>
@@ -120,6 +120,7 @@ class SimpleView extends React.Component {
                   </div>
                 </div>
               }
+              <ShareReport/>
             </div>
             <div className="clearfix" />
           </div>
@@ -138,9 +139,6 @@ class SimpleView extends React.Component {
         </LabeledBox>
         <LabeledBox label={<TextMaker text_key="blue_text_report_data_sources" args={{table_name: table.name}} />}>
           <ReportDatasets {...this.props} /> 
-        </LabeledBox>
-        <LabeledBox label={<TextMaker text_key="share_report"/>}>
-          <ShareReport /> 
         </LabeledBox>
         { table.rpb_banner && <AlertBanner>{table.rpb_banner}</AlertBanner> }
         <div id="rpb-main-content" >
